Finish first pass rendering workflow results
