+++ awx/ui_next/src/components/Workflow/index.js
@@ -1,0 +1,3 @@
+export { default as WorkflowHelp } from './WorkflowHelp';
+export { default as WorkflowHelpDetails } from './WorkflowHelpDetails';
+export { default as WorkflowWrapper } from './WorkflowWrapper';
