+++ awx/ui_next/src/screens/Job/WorkflowOutput/index.js
@@ -1,1 +1,2 @@
-export { default } from './WorkflowOutput';
+export { default as WorkflowOutput } from './WorkflowOutput';
+export { default as WorkflowOutputGraph } from './WorkflowOutputGraph';
