+++ awx/ui_next/src/util/dates.jsx
@@ -4,3 +4,7 @@ import { getLanguage } from './language';
 export function formatDateString(dateString, lang = getLanguage(navigator)) {
   return new Date(dateString).toLocaleString(lang);
 }
+
+export function secondsToHHMMSS(seconds) {
+  return new Date(seconds * 1000).toISOString().substr(11, 8);
+}
